refactor: move resolveLocallyFirst to build
